update quantity field on PackCommand

+++ grails-app/controllers/uk/co/wonderlane/wlpos/ProductController.groovy
@@ -1862,7 +1862,7 @@ class ProductVariantCommand {
 class PackCommand {
     int id
     Supplier supplier
-    int quantity
+    BigDecimal quantity
     BigDecimal price
     String orderCode
     String barcode
